UX-493 Move source code to ui frame
